FFmpeg processor & updated .gitlab-ci.yml
All FFmpeg classes now inherit from "FFmpeg" abstract class, which makes it simpler to do not have duplicated code and bug tracking. In addition, when the YouTube data is downloaded, by using FFmpeg, the program removes the video, if exists, and normalizes the audio. If tests go well, we can mark audio milestone as completed.