Updated tests and close method
Trying to do close operations inside the "__del__" method is dangerous as some items may be garbage collected, so this can lead into deadlocks and race conditions, even class not being properly deleted. This way, the deletion method was moved to "close" or "stop" method and "__del__" tries to call them safely wrapping inside a try/catch block. This behaviour may be removed in a future, making necessary to call the proper close method raising an exception if it was not called